+++ core-api/src/main/java/com/optimizely/ab/decision/evaluator/AudienceEvaluator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and      *
  * limitations under the License.                                           *
  ***************************************************************************/
-package com.optimizely.ab.decision.audience;
+package com.optimizely.ab.decision.evaluator;
 
 import com.optimizely.ab.config.Experiment;
 import com.optimizely.ab.config.ProjectConfig;
